Move Lrama::Rule to Lrama::Grammar::Rule
