+++ lib/lrama.rb
@@ -9,7 +9,6 @@ require "lrama/output"
 require "lrama/parser"
 require "lrama/reference"
 require "lrama/report"
-require "lrama/rule"
 require "lrama/state"
 require "lrama/states"
 require "lrama/states_reporter"
